+++ src/main/java/teammates/logic/api/Logic.java
@@ -1341,7 +1341,7 @@ public class Logic {
     /**
      * Saves a feedback response record.
      */
-    public void createFeedbackResponseRecord(int count, int timestamp)
+    public void createFeedbackResponseRecord(long count, long timestamp)
             throws InvalidParametersException, EntityAlreadyExistsException {
         feedbackResponseRecordLogic.createFeedbackResponseRecord(count, timestamp);
     }
